feat: extract gateway_posts_rest and mount via capability registry
Move GET /posts/{post_id} into a new gateway_posts_rest package, mounted
through the capability registry like gateway_notes_rest and
gateway_following_rest. GET /posts/{post_id}/comments stays in
gateway_oss/api/v1/posts.py for now (moves to gateway_comments_rest in a
later task).
