+++ packages/gateway_oss/pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
   "gateway_notes_mcp",
   "gateway_following_rest",
   "gateway_following_mcp",
+  "gateway_posts_rest",
   "gateway_rest_common",
   "gateway_mcp_common",
 ]
@@ -41,5 +42,6 @@ gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
 gateway_following_mcp = { workspace = true }
+gateway_posts_rest = { workspace = true }
 gateway_rest_common = { workspace = true }
 gateway_mcp_common = { workspace = true }
