+++ packages/gateway_oss/src/gateway_oss/api/v1/posts.py
@@ -5,20 +5,10 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, Path
 
 from gateway_oss.api.deps import get_posts_service
-from gateway_oss.models.schemas import CommentsResponse, FullPostResponse
+from gateway_oss.models.schemas import CommentsResponse
 from gateway_oss.services.posts import PostsService
 
 router = APIRouter(tags=["posts"])
-
-
-@router.get("/posts/{post_id}", response_model=FullPostResponse)
-async def get_post(
-    post_id: Annotated[int, Path(gt=0)],
-    service: Annotated[PostsService, Depends(get_posts_service)],
-) -> FullPostResponse:
-    """Return a single Substack post with its full content."""
-    post = await service.get_post_by_id(post_id)
-    return FullPostResponse.from_substack(post)
 
 
 @router.get("/posts/{post_id}/comments", response_model=CommentsResponse)
