+++ packages/gateway_posts_rest/pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "gateway_posts_rest"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_rest_common", "gateway_posts", "fastapi>=0.138.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+posts_rest = "gateway_posts_rest:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_posts_rest"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_rest_common = { workspace = true }
+gateway_posts = { workspace = true }
