+++ pyproject.toml
@@ -32,6 +32,7 @@ gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
 gateway_following_mcp = { workspace = true }
+gateway_posts_rest = { workspace = true }
 gateway_rest_common = { workspace = true }
 gateway_mcp_common = { workspace = true }
 
@@ -45,7 +46,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_posts_rest", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
